+++ src/utils/error/errors.rs
@@ -1,19 +1,35 @@
 #[path = "../ui/color_texts.rs"]
 mod font_styles;
 
+#[allow(dead_code)]
 pub enum Errors {
-    SyntaxError,
+    _SyntaxError,
+    _InputError,
+    _DatabaseBoolError,
 }
 
 impl Errors {
     // Public function
     // Throws an error
-    pub fn throw_error(&self) {
+    pub fn _throw_error(&self) {
         match *self {
-            Errors::SyntaxError => font_styles::FontStyle::_error_message(
-                &font_styles::FontStyle::Red,
-                String::from("SyntaxError"),
-                String::from("An error occured since command given doesn't exist"),
+            Errors::_SyntaxError => {
+                let code = "todo --help";
+                font_styles::_error_message(
+                    String::from("SyntaxError"),
+                    String::from("An error occured since command given doesn't exist"),
+                );
+                println!("❓ Try this : \x1b[1;3;35m{}\x1b[0m", code);
+            }
+
+            Errors::_InputError => font_styles::_error_message(
+                String::from("InputError"),
+                String::from("An error occured since no input was given, please try again!"),
+            ),
+
+            Errors::_DatabaseBoolError => font_styles::_error_message(
+                String::from("DatabaseBoolError"),
+                String::from("An error occured since no boolean value in database was corrupted!"),
             ),
         }
     }